修复祭祀回复逻辑和评分规则

+++ Bloodhaver.cpp
@@ -11,7 +11,7 @@ int BloodHaver::getDEF(int attackType_got)
 }
 
 
-void BloodHaver::updateBlood(int damage){
+void BloodHaver::updateBlood(double damage){
     double maxBlood=(double)getMaxBlood();
-    Blood -= (double)damage/maxBlood; Blood=max(min(1.0,Blood),0.0);
+    Blood -= damage/maxBlood; Blood=max(min(1.0,Blood),0.0);
 }
